requirements.txt
moviepy==1.0.3
pillow==10.0.0
requests==2.31.0 